Moved the landingpage into a pages directory and updated the readme
